+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v 0.2
+- Added `/imagegen` command
+- Rate limits configurable per command
+
 ## v 0.1.1
 - Media directory is configurable using the `KEWLBOT_MEDIA_FOLDER` environment variable